+++ src/components/manifold-data-rename-button/manifold-data-rename-button.tsx
@@ -45,6 +45,7 @@ export class ManifoldDataRenameButton {
   /** The id of the resource to rename, will be fetched if not set */
   @Prop({ mutable: true }) resourceId?: string = '';
   @Prop() loading?: boolean = false;
+  @Prop() disabled?: boolean;
   @Event({ eventName: 'manifold-renameButton-click', bubbles: true }) click: EventEmitter;
   @Event({ eventName: 'manifold-renameButton-invalid', bubbles: true }) invalid: EventEmitter;
   @Event({ eventName: 'manifold-renameButton-error', bubbles: true }) error: EventEmitter;
@@ -188,7 +189,7 @@ export class ManifoldDataRenameButton {
       <button
         type="submit"
         onClick={() => this.rename()}
-        disabled={!this.resourceId && !this.loading}
+        disabled={(!this.resourceId && !this.loading) || this.disabled}
       >
         <slot />
       </button>